Improve logging system
